Recipie_App/requirements.txt
pymongo 
pillow 